Собственный класс исключения

+++ opinions_app/api_views.py
@@ -1,31 +1,48 @@
+# what_to_watch/opinions_app/api_views.py
+
 from flask import jsonify, request
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
+
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     return jsonify({'opinion': opinion.to_dict()}), 200
+
 
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
-    data = request.json()
-    opinion = Opinion.query.get_or_404(id)
-    if Opinion.query.filter_by(text=data['text']).first is not None:
-        return jsonify({'error': 'Такое мнение в базе уже есть'}), 400
-
-    opinion.title = request.json.get('title', opinion.title)
-    opinion.text = request.json.get('text', opinion.text)
-    opinion.source = request.json.get('source', opinion.source)
-    opinion.added_by = request.json.get('added_by', opinion.added_by)
+    data = request.get_json()
+    if (
+        'text' in data and 
+        Opinion.query.filter_by(text=data['text']).first() is not None
+    ):
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
+    opinion = Opinion.query.get(id)
+    # Тут код ответа нужно указать явным образом.
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
+    opinion.title = data.get('title', opinion.title)
+    opinion.text = data.get('text', opinion.text)
+    opinion.source = data.get('source', opinion.source)
+    opinion.added_by = data.get('added_by', opinion.added_by)
     db.session.commit()  
     return jsonify({'opinion': opinion.to_dict()}), 201
 
+
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -40,21 +57,22 @@ def get_opinions():
 
 @app.route('/api/opinions/', methods=['POST'])
 def add_opinion():
-    data = request.get_json(silent=True)
-    reqired_fields = ['title', 'text']
-    if data is None or set(reqired_fields) & set(data.keys()) != set(reqired_fields):
-        return jsonify({'error': 'В запросе отсутствуют обязательные поля'}), 400
-    
-    if Opinion.query.filter_by(text=data['text']).first is not None:
-        return jsonify({'error': 'Такое мнение в базе уже есть'}), 400
-
+    data = request.get_json()
+    if 'title' not in data or 'text' not in data:
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
+    if Opinion.query.filter_by(text=data['text']).first() is not None:
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     opinion = Opinion()
     opinion.from_dict(data)
     db.session.add(opinion)
     db.session.commit()
     return jsonify({'opinion': opinion.to_dict()}), 201
 
+
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    # Тут код ответа нужно указать явным образом.
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
